pyproject.toml
[tool.poetry]
name = "instagram_django"
version = "0.1.0"
description = ""
authors = ["Jaesik <joje1283@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.2"
gunicorn = {version = "^20.1.0", extras = ["production"]}
uvicorn = {extras = ["production"], version = "^0.17.4"}
djangorestframework = "^3.13.1"
Pillow = "^9.0.1"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
